# openapi.yaml
openapi: 3.0.0

info:
  title: Mood Joke Generator API
  version: 2.0.0
  description: |
    An agentic joke API powered by Claude. Submit a mood in plain natural language
    and an AI agent will interpret it, query the joke datastore, and return the
    most relevant joke. Rate jokes to build an aggregate feedback dataset.

    ## How the agent works
    `POST /get-joke` does not do a simple keyword lookup. The raw mood string is
    passed to a Claude agent that:
    1. Interprets the emotional context using natural language understanding
    2. Chooses its own short category label for the mood (no fixed list)
    3. Calls an internal `pick_jokes` tool to retrieve candidate jokes from the
       configured datastore (flat text file or vector database)
    4. Selects and returns the single most appropriate joke

    ## Datastore backends
    The active backend is controlled by the `STORE` environment variable on the
    server:
    - `local` — reads from a flat `randomJokes.txt` file, one joke per line (default)
    - `vector` — queries a vector database for semantic similarity search

  contact:
    name: Mood Joke Generator
  license:
    name: MIT

servers:
  - url: http://localhost:5500
    description: Local development server

tags:
  - name: Jokes
    description: Retrieve AI-curated jokes based on natural language mood input
  - name: Ratings
    description: Submit and retrieve joke ratings
  - name: Stats
    description: Aggregate rating statistics

paths:

  # ─── POST /get-joke ────────────────────────────────────────────────────────
  /get-joke:
    post:
      operationId: getJoke
      tags: [Jokes]
      summary: Get a joke matched to a mood
      description: |
        Accepts a free-text mood description and returns the most contextually
        appropriate joke from the datastore, selected by an AI agent.

        The `mood` field accepts any natural language description — the agent
        invents a category label on the fly. Examples:
        - `"happy"` — agent picks a category like `happy`
        - `"feeling down after a rough day"` — agent picks something like `sad`
        - `"completely wiped out from jet lag"` — agent picks something like `tired`
        - `"my presentation just went sideways"` — agent picks something like `stressed`

        The agent always calls the internal `pick_jokes` tool before selecting a
        joke, so candidates are drawn from the real datastore rather than
        generated by the LLM.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetJokeRequest'
            examples:
              simple_mood:
                summary: Exact category keyword
                value:
                  mood: happy
              natural_language_tired:
                summary: Natural language — tired
                value:
                  mood: completely wiped out from jet lag
              natural_language_stressed:
                summary: Natural language — stressed
                value:
                  mood: my presentation just went sideways and I want to hide
              natural_language_sad:
                summary: Natural language — sad
                value:
                  mood: feeling really down after a rough week
      responses:
        '200':
          description: A joke was successfully selected by the agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetJokeResponse'
              examples:
                success:
                  summary: Successful joke response
                  value:
                    joke: Why don't scientists trust atoms? Because they make up everything!
                    mood: happy
                natural_language:
                  summary: Natural language mood mapped by the agent
                  value:
                    joke: I stayed up all night to see where the sun went. Then it dawned on me.
                    mood: completely wiped out from jet lag
        '400':
          description: The `mood` field was missing, empty, or whitespace-only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_mood:
                  summary: Missing mood field
                  value:
                    error: mood is required
                empty_mood:
                  summary: Empty string
                  value:
                    error: mood is required
        '500':
          description: The AI agent encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                agent_error:
                  summary: Agent failure
                  value:
                    error: The joke agent encountered an error. Please try again.

  # ─── POST /rate-joke ───────────────────────────────────────────────────────
  /rate-joke:
    post:
      operationId: rateJoke
      tags: [Ratings]
      summary: Submit a rating for a joke
      description: |
        Records a 1–3 star rating for a previously returned joke. The `mood` and
        `joke` fields are echoed back from the `/get-joke` response so each
        rating is linked to its context.

        Ratings are stored in memory on the server and can be retrieved via
        `GET /stats`. The rating scale is:
        - `1` — Meh (😐)
        - `2` — Good (🙂)
        - `3` — Hilarious (😂)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateJokeRequest'
            examples:
              hilarious:
                summary: Highest rating
                value:
                  mood: happy
                  joke: Why don't scientists trust atoms? Because they make up everything!
                  rating: 3
              meh:
                summary: Lowest rating
                value:
                  mood: tired
                  joke: I'm reading a book on anti-gravity. It's impossible to put down!
                  rating: 1
      responses:
        '200':
          description: Rating recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateJokeResponse'
              examples:
                success:
                  value:
                    message: Thanks for your feedback!

  # ─── GET /stats ────────────────────────────────────────────────────────────
  /stats:
    get:
      operationId: getStats
      tags: [Stats]
      summary: Get aggregate rating statistics
      description: |
        Returns the total number of ratings submitted, the average rating score
        across all jokes, and the full log of individual rating entries.

        `averageRating` is returned as a string formatted to 2 decimal places
        (e.g. `"2.33"`), or `null` when no ratings have been submitted yet.

        Note: ratings are stored in server memory and reset on server restart.
      responses:
        '200':
          description: Aggregate statistics and full rating log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
              examples:
                with_ratings:
                  summary: Stats with existing ratings
                  value:
                    totalRatings: 3
                    averageRating: "2.33"
                    ratings:
                      - mood: happy
                        joke: Why don't scientists trust atoms? Because they make up everything!
                        rating: 3
                        time: "2025-06-01T10:00:00.000Z"
                      - mood: tired
                        joke: I'm reading a book on anti-gravity. It's impossible to put down!
                        rating: 2
                        time: "2025-06-01T10:05:00.000Z"
                      - mood: sad
                        joke: Why don't skeletons fight each other? They don't have the guts.
                        rating: 2
                        time: "2025-06-01T10:10:00.000Z"
                no_ratings:
                  summary: No ratings submitted yet
                  value:
                    totalRatings: 0
                    averageRating: null
                    ratings: []

# ─── Components ───────────────────────────────────────────────────────────────
components:
  schemas:

    # ── Requests ──────────────────────────────────────────────────────────────

    GetJokeRequest:
      type: object
      required: [mood]
      properties:
        mood:
          type: string
          minLength: 1
          description: |
            A free-text description of the user's current emotional state.
            Accepts any natural language — the AI agent will interpret it and
            map it to the most relevant jokes in the datastore.
          examples:
            - happy
            - feeling really down after a tough week
            - completely wired and can't sit still
            - wiped out after a red-eye flight

    RateJokeRequest:
      type: object
      required: [mood, joke, rating]
      properties:
        mood:
          type: string
          minLength: 1
          description: The mood string originally sent to `POST /get-joke`
          example: happy
        joke:
          type: string
          minLength: 1
          description: The joke text returned by `POST /get-joke`
          example: Why don't scientists trust atoms? Because they make up everything!
        rating:
          oneOf:
            - type: integer
            - type: string
          description: |
            Rating value on a 1–3 scale. Accepts either an integer or a
            numeric string — the server coerces it to a number.
            - `1` — Meh
            - `2` — Good
            - `3` — Hilarious
          enum: [1, 2, 3, "1", "2", "3"]
          example: 3

    # ── Responses ─────────────────────────────────────────────────────────────

    GetJokeResponse:
      type: object
      required: [joke, mood]
      properties:
        joke:
          type: string
          description: The joke selected by the AI agent from the datastore
          example: Why don't scientists trust atoms? Because they make up everything!
        mood:
          type: string
          description: The raw mood string that was submitted in the request
          example: happy

    RateJokeResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Confirmation message
          example: Thanks for your feedback!

    StatsResponse:
      type: object
      required: [totalRatings, averageRating, ratings]
      properties:
        totalRatings:
          type: integer
          minimum: 0
          description: Total number of ratings recorded since the server started
          example: 42
        averageRating:
          type:
            - string
            - "null"
          description: |
            Mean rating across all entries, formatted to 2 decimal places.
            Returns `null` when no ratings have been submitted.
          example: "2.33"
        ratings:
          type: array
          description: Full log of individual rating entries
          items:
            $ref: '#/components/schemas/RatingEntry'

    RatingEntry:
      type: object
      required: [mood, joke, rating, time]
      properties:
        mood:
          type: string
          description: The mood submitted with the original joke request
          example: happy
        joke:
          type: string
          description: The joke text that was rated
          example: Why don't scientists trust atoms? Because they make up everything!
        rating:
          type: integer
          minimum: 1
          maximum: 3
          description: The submitted rating (1 = Meh, 2 = Good, 3 = Hilarious)
          example: 3
        time:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the rating was recorded
          example: "2025-06-01T10:00:00.000Z"