# server.js
require("dotenv").config();
const fs              = require("fs");
const express         = require("express");
const path            = require("path");
const yaml            = require("js-yaml");
const swaggerJSDoc    = require("swagger-jsdoc");
const swaggerUi       = require("swagger-ui-express");
const { createStore } = require("./stores");
const { runJokeAgent } = require("./agent/jokeAgent");
const { createProvider } = require("./providers");
const cors            = require("cors");

const app  = express();
const PORT = process.env.PORT || 3000;

/* ── Initialise the joke store (local file or vector DB) ─────────────────── */
const store = createStore();
const provider = createProvider();

/* Hot-reload the local store on SIGHUP (kill -HUP <pid>) */
process.on("SIGHUP", () => {
  if (typeof store.reload === "function") {
    console.log("🔄  Reloading joke store...");
    store.reload();
  }
});

app.use(cors({
  origin: process.env.FRONTEND_URL || "http://localhost:8080",
  methods: ["GET", "POST"],
}));

app.use(express.json());
//app.use(express.static(path.join(__dirname, "public")));

let ratings = [];

/* ── Swagger / OpenAPI ────────────────────────────────────────────────────
   openapi.yaml supplies the base spec (info, servers, tags, components).
   Per-endpoint `@swagger` JSDoc blocks below contribute the `paths` section.
   Mounted at /docs (interactive UI) and /openapi.json (raw spec).
─────────────────────────────────────────────────────────────────────────── */
const baseSpec = yaml.load(fs.readFileSync(path.join(__dirname, "openapi.yaml"), "utf8"));
delete baseSpec.paths; // paths live in @swagger JSDoc — keep one source of truth

const swaggerSpec = swaggerJSDoc({
  definition: baseSpec,
  apis: [__filename],
});

app.use(
  "/docs",
  swaggerUi.serve,
  swaggerUi.setup(swaggerSpec, {
    customSiteTitle: "Mood Joke Generator API",
    swaggerOptions:  { docExpansion: "list", defaultModelsExpandDepth: 0 },
  })
);
app.get("/openapi.json", (_req, res) => res.json(swaggerSpec));

/* ── Routes ──────────────────────────────────────────────────────────────── */

/**
 * @swagger
 * /get-joke:
 *   post:
 *     operationId: getJoke
 *     tags: [Jokes]
 *     summary: Get a joke matched to a mood
 *     description: |
 *       Accepts a free-text mood description and returns the most contextually
 *       appropriate joke from the datastore, selected by an AI agent.
 *
 *       The `mood` field accepts any natural language description — the agent
 *       invents a category label on the fly. Examples:
 *       - `"happy"` — agent picks a category like `happy`
 *       - `"feeling down after a rough day"` — agent picks something like `sad`
 *       - `"completely wiped out from jet lag"` — agent picks something like `tired`
 *       - `"my presentation just went sideways"` — agent picks something like `stressed`
 *
 *       The agent always calls the internal `pick_jokes` tool before selecting a
 *       joke, so candidates are drawn from the real datastore rather than
 *       generated by the LLM.
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/GetJokeRequest'
 *           examples:
 *             simple_mood:
 *               summary: Exact category keyword
 *               value: { mood: happy }
 *             natural_language_tired:
 *               summary: Natural language — tired
 *               value: { mood: completely wiped out from jet lag }
 *             natural_language_stressed:
 *               summary: Natural language — stressed
 *               value: { mood: my presentation just went sideways and I want to hide }
 *             natural_language_sad:
 *               summary: Natural language — sad
 *               value: { mood: feeling really down after a rough week }
 *     responses:
 *       '200':
 *         description: A joke was successfully selected by the agent
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/GetJokeResponse'
 *             examples:
 *               success:
 *                 summary: Successful joke response
 *                 value:
 *                   joke: Why don't scientists trust atoms? Because they make up everything!
 *                   mood: happy
 *               natural_language:
 *                 summary: Natural language mood mapped by the agent
 *                 value:
 *                   joke: I stayed up all night to see where the sun went. Then it dawned on me.
 *                   mood: completely wiped out from jet lag
 *       '400':
 *         description: The `mood` field was missing, empty, or whitespace-only
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *             examples:
 *               missing_mood:
 *                 summary: Missing mood field
 *                 value: { error: mood is required }
 *               empty_mood:
 *                 summary: Empty string
 *                 value: { error: mood is required }
 *       '500':
 *         description: The AI agent encountered an internal error
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/ErrorResponse'
 *             examples:
 *               agent_error:
 *                 summary: Agent failure
 *                 value: { error: The joke agent encountered an error. Please try again. }
 */
app.post("/get-joke", async (req, res) => {
  const mood = (req.body.mood || "").trim();

  if (!mood) {
    return res.status(400).json({ error: "mood is required" });
  }

  try {
    console.log(`😊  /get-joke mood="${mood}"`);
    const joke = await runJokeAgent(mood, store, provider);
    res.json({ joke, mood });
  } catch (err) {
    console.error("❌  Agent error:", err.message);
    res.status(500).json({ error: "The joke agent encountered an error. Please try again." });
  }
});

/**
 * @swagger
 * /rate-joke:
 *   post:
 *     operationId: rateJoke
 *     tags: [Ratings]
 *     summary: Submit a rating for a joke
 *     description: |
 *       Records a 1–3 star rating for a previously returned joke. The `mood` and
 *       `joke` fields are echoed back from the `/get-joke` response so each
 *       rating is linked to its context.
 *
 *       Ratings are stored in memory on the server and can be retrieved via
 *       `GET /stats`. The rating scale is:
 *       - `1` — Meh (😐)
 *       - `2` — Good (🙂)
 *       - `3` — Hilarious (😂)
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/RateJokeRequest'
 *           examples:
 *             hilarious:
 *               summary: Highest rating
 *               value:
 *                 mood: happy
 *                 joke: Why don't scientists trust atoms? Because they make up everything!
 *                 rating: 3
 *             meh:
 *               summary: Lowest rating
 *               value:
 *                 mood: tired
 *                 joke: I'm reading a book on anti-gravity. It's impossible to put down!
 *                 rating: 1
 *     responses:
 *       '200':
 *         description: Rating recorded successfully
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/RateJokeResponse'
 *             examples:
 *               success:
 *                 value: { message: Thanks for your feedback! }
 */
app.post("/rate-joke", (req, res) => {
  const { mood, joke, rating } = req.body;
  ratings.push({ mood, joke, rating: Number(rating), time: new Date() });
  console.log(`⭐  /rate-joke rating=${rating} mood="${mood}"`);
  res.json({ message: "Thanks for your feedback!" });
});

/**
 * @swagger
 * /stats:
 *   get:
 *     operationId: getStats
 *     tags: [Stats]
 *     summary: Get aggregate rating statistics
 *     description: |
 *       Returns the total number of ratings submitted, the average rating score
 *       across all jokes, and the full log of individual rating entries.
 *
 *       `averageRating` is returned as a string formatted to 2 decimal places
 *       (e.g. `"2.33"`), or `null` when no ratings have been submitted yet.
 *
 *       Note: ratings are stored in server memory and reset on server restart.
 *     responses:
 *       '200':
 *         description: Aggregate statistics and full rating log
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/StatsResponse'
 *             examples:
 *               with_ratings:
 *                 summary: Stats with existing ratings
 *                 value:
 *                   totalRatings: 3
 *                   averageRating: "2.33"
 *                   ratings:
 *                     - { mood: happy, joke: "Why don't scientists trust atoms? Because they make up everything!", rating: 3, time: "2025-06-01T10:00:00.000Z" }
 *                     - { mood: tired, joke: "I'm reading a book on anti-gravity. It's impossible to put down!",  rating: 2, time: "2025-06-01T10:05:00.000Z" }
 *                     - { mood: sad,   joke: "Why don't skeletons fight each other? They don't have the guts.",   rating: 2, time: "2025-06-01T10:10:00.000Z" }
 *               no_ratings:
 *                 summary: No ratings submitted yet
 *                 value:
 *                   totalRatings: 0
 *                   averageRating: null
 *                   ratings: []
 */
app.get("/stats", (req, res) => {
  const avg = ratings.length
    ? (ratings.reduce((s, r) => s + r.rating, 0) / ratings.length).toFixed(2)
    : null;
  res.json({ totalRatings: ratings.length, averageRating: avg, ratings });
});

app.listen(PORT, () => {
  const storeType = (process.env.STORE || "local").toUpperCase();
  console.log(`✅  Mood Joke Generator running → http://localhost:${PORT}`);
  console.log(`📦  Store: ${storeType}`);
  console.log(`📘  API Docs:    http://localhost:${PORT}/docs`);
});